Extract States page to component
